requirements/base.txt
# Core Django
Django==6.0.2
asgiref==3.11.1
sqlparse==0.5.5
tzdata==2025.3

# Environment / configuration
django-environ==0.12.0

# Database
psycopg[binary]==3.2.9

# Server
gunicorn==23.0.0

# Static files
whitenoise[brotli]==6.9.0

# Security
django-cors-headers==4.7.0
